Update PsMapExec.ps1
Improve SMB Performance: Updated the way remote privileges are confirmed on target systems with the SMB protocol. Testing this with 500x computers now shows a 10x increase in speed!

SessionHunter: Improved filtering on reachable hosts. Improves the overall speed of the script but still need a bit of work to run in parallel. 

GenRelayList: Improved the speed of the port checking and the error handling. Altogether this method should now be much faster.